feat(sdk): Implement SDK client with model listing and content generation functionality
It's a very simple client I could generate it using openapitools but I saw the version of the generated code was 1.18 so I've decided to create the simple SDK until I find a better solution to generate this client. It's not a lot of code because there are only 2 API endpoints.

Signed-off-by: Eden Reich <eden.reich@gmail.com>