docs(openapi): Add CreateCompletion endpoint and request/response schemas
Seems like the older version of this API, but some coding tools still using it, so adding the minimal of it for backward compatibility.

I assume it's like in Ollama API where you send tokens and you receive some tokens back - only the completion rather than the entire chat. Perhaps it's there for a reason because it's more efficient, less tokens needs to be passed over the connection wire. Curious to see how the IDE's is using it.

Signed-off-by: Eden Reich <eden.reich@gmail.com>