refactor(api): Simplify streaming API
Removed some unnecessary logic, since most providers have now OpenAI compatibility API that means no extra logic needs to be implemented, sending the response back to the client as-is, which results in better performance.

Still need to test Ollama, when the server is timing out the streaming is being closed unexpectedly, I didn't figure out how to increase the timeout on Ollama server, since it's running locally it takes longer to generate the tokens. I did find out in their documentation how to enable Debugging messages, and it's a bit clearer that my request was failing due to timeout on Ollama server. It's nowhere documented so I'll have to figure out.

Other providers works almost seamlessly. Need to write new tests with this OpenAI compatible API.

Signed-off-by: Eden Reich <eden.reich@gmail.com>