fix(middleware): Add streaming request handling and context key for SSE events
Sometimes the client might send a regular HTTP request, so they will not see the SSE for the progressing tool_calls.

There is work around it, if I introduce client notifications to a webhook, but I will keep it out of the scope of this feature pull request because on the UI I will mainly just use streaming.

Signed-off-by: Eden Reich <eden.reich@gmail.com>