fix: MCP client initialization and better warning feedback
When no MCP servers are configured and the middleware was enabled the user should see what's happening in the logs and understand it. Right now it was throwing a nil pointer error. Because the client was initialized as nil and then used as a middleware, instead we'll use the no op middleware.

Signed-off-by: Eden Reich <eden.reich@gmail.com>