refactor(agent): Update completion signal to [DONE] and enhance chunk processing logic
Simplify it, I though I would need an extra [AGENT_DONE] but actually I could just filter out the [DONE] from the subsequent requests to provider and write the [DONE] when the agent actually done.

TODO - I still need to refactor the part where I parse the data, it's seems a bit clanky. I tested it with Groq and DeepSeek and I had to add this part when using DeepSeek not entirely sure yet why, but I'll investigate.

Signed-off-by: Eden Reich <eden.reich@gmail.com>