refactor: Refactor A2A Agent and Client Implementation
Instead of adding the skills as tools and coupling the a2a-compatible agents with the inference gateway, the a2a-compatible agent should not care who submits the task it only care that it receives a task.

The agent is the abstractions of the skills. the skills are processed within the agent. I initially implement it wrong and this is now working way better.

Now, the inference gateway a2a-compatible agent is only able to query the card of a specific agent and submit a task - that's all.

Signed-off-by: Eden Reich <eden.reich@gmail.com>